requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.2.1
django-cors-headers==4.7.0
django-ninja==1.4.1
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
python-decouple==3.8
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
whitenoise==6.9.0
